outer plugins: rename the .scm files and the plugin procedures to standard names
plugin => plug-in
script-fu- => plug-in-

The standard for names in the PDB is that plug-in is hyphenated.
In natural language, I don't hyphenate.
And I disagree that it should be hyphenated in the code.
But until the standard is changed, use hyphenation
so the set of names in the PDB is consistent.

Plugins that are served by extension-script-fu and installed in /scripts
are named script-fu-<foo>.
While plugins independendently interpreted and installed to /plug-ins
are named plug-in-<foo>.

Miscellaneous changes to plugin that tests resynthesizer, a Python file.