Fix engine handling of indexed mode images.
In v2, engine and outer plugins not work with indexed mode images.
There is no technical reason they shouldn't.
It might be rare a user wants to use indexed images,
and the results are not good if the palette is not ordered,
but allow user to do it and judge for themselves.

Some work to enable this was done in earlier commits.

Fix the Babl format of indexed image: Y' u8 => Y u8
I.E. indexed is "linear" not non-linear.
Avoid any conversions that might apply a non-linear converstion function.

Add tests on indexed mode images.