Support building on macOS.
The GIMP App on macOS has all the shared libraries that the plugin
needs, it just doesn't have the headers. This change uses include
paths from pre-built libraries (from Homebrew or MacPorts) and the
GIMP source tree (from git or a release tarball) to build the plugin
and force it to use the shared libraries that are installed with
GIMP.app.